[repo] Move `decimojo` folder into `src` folder (#17)
This pull request includes several changes to the `mojoproject.toml` and
`src/decimojo/decimal.mojo` files. The most important changes involve
moving `decimal` foder into `src` folder, fixing name of the project in
toml, updating the packaging command, and improving the documentation
for the internal representation of the `Decimal` struct.

Project renaming and packaging:

*
[`mojoproject.toml`](diffhunk://#diff-1b0ef62120bccf4c05a76e60e13fe686f33eb92d5897b3f05e0c3f8d737fc5c0L6-R6):
Fix the project name from `decimal` to `decimojo` and updated the
packaging command to use the `src/decimojo` directory.
[[1]](diffhunk://#diff-1b0ef62120bccf4c05a76e60e13fe686f33eb92d5897b3f05e0c3f8d737fc5c0L6-R6)
[[2]](diffhunk://#diff-1b0ef62120bccf4c05a76e60e13fe686f33eb92d5897b3f05e0c3f8d737fc5c0L16-R16)

Documentation improvements:

*
[`src/decimojo/decimal.mojo`](diffhunk://#diff-c2178e6e0bbbbb0d3379f1513412224309d1232c444654e77abb5ce36b3a854fL42-R47):
Enhanced the documentation for the internal representation of the
`Decimal` struct, providing clearer descriptions of the bit fields and
their roles.
[[1]](diffhunk://#diff-c2178e6e0bbbbb0d3379f1513412224309d1232c444654e77abb5ce36b3a854fL42-R47)
[[2]](diffhunk://#diff-c2178e6e0bbbbb0d3379f1513412224309d1232c444654e77abb5ce36b3a854fL54-R57)