[decimal] Implement the `log()` and the `log10()` functions (#43)
This pull request includes the `log` and `log10` functions in the
`decimojo` library.
It also introduces a new benchmark for the `log10` function and includes
several updates to the `Decimal` struct in the `decimojo` library.

### Library Updates:
* Included `log` and `log10` in the `decimojo` library's `__init__.mojo`
file to make them available for import.

### New Benchmark for `log10` Function:
* Added a comprehensive benchmark for the `log10` function in
`benches/bench_log10.mojo` to compare performance against Python's
decimal module with diverse test cases.

### Project Configuration Updates:
* Updated `benches/bench.mojo` to include `bench_log10` in the list of
benchmarks.
* Updated `mojoproject.toml` to include `test_log` and `test_log10` in
the test configurations and `bench_log10` in the benchmark
configurations.

### `Decimal` Struct Enhancements:
* Added `@always_inline` decorator to several static methods and
instance methods in `struct Decimal` to improve performance.