[doc] Update changelog and readme files (#62)
This pull request includes several updates and improvements to the
`README.md` and other documentation files for the DeciMojo library, as
well as the addition of a new changelog. The most important changes
include the introduction of `BigUInt`, updates to the installation
instructions, and enhancements to the `BigInt` and `Decimal`
functionalities.

### Enhancements to `BigInt` and `BigUInt`:

* Added a new `BigUInt` type and extended the `BigInt` type to support
unlimited precision arithmetic, comprehensive arithmetic operations, and
efficient handling of extremely large integers.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L14-R14)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R178)
* Enhanced `BigInt` with floor division, truncate division, modulo, and
power operations.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L114-R122)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R136-R139)

### Updates to Installation Instructions:

* Updated the installation instructions to reflect the new version
`0.2.0` in the `mojoproject.toml` file.

### Documentation Improvements:

* Updated the `README.md` to reflect changes in the `BigInt` and
`Decimal` types, including new mathematical functions and division
operations.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L103-R103)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L215-R228)
* Added a new `docs/changelog.md` file to document the changes and
improvements in version `0.2.0`, including performance benchmarks and
detailed descriptions of new features and bug fixes.

### Additional Changes:

* Made minor updates to the Chinese documentation (`docs/readme_zht.md`)
to include the new `BigUInt` type and updated installation instructions.