[decimal] Implement `root()` for `BigDecimal` (#80)
This pull request introduces a new root function for the BigDecimal
library and makes several updates to the benchmarking and testing code
to support this new functionality. Additionally, some improvements and
optimizations have been made to existing functions.

Key changes include:

### New Functionality:
* Added a new `root` function to the `BigDecimal` struct for calculating
the nth root of a BigDecimal number.
* Introduced `integer_root` and `is_integer_reciprocal_and_return`
helper functions in `exponential.mojo` to support the new root
calculation.

### Benchmarking Updates:
* Updated `benches/bigdecimal/bench.mojo` to include the new `root`
benchmark, adding the necessary imports, command handling, and execution
logic.
[[1]](diffhunk://#diff-96175f1000f81fe074f70a9adc4b39cd3f2dac8e85d30aa7efed9fc02b46e5ecR8)
[[2]](diffhunk://#diff-96175f1000f81fe074f70a9adc4b39cd3f2dac8e85d30aa7efed9fc02b46e5ecR25)
[[3]](diffhunk://#diff-96175f1000f81fe074f70a9adc4b39cd3f2dac8e85d30aa7efed9fc02b46e5ecR48-R49)
[[4]](diffhunk://#diff-96175f1000f81fe074f70a9adc4b39cd3f2dac8e85d30aa7efed9fc02b46e5ecR58)

### Precision Handling:
* Modified the `sqrt` function calls in `bench_bigdecimal_sqrt.mojo` and
`test_bigdecimal_exponential.mojo` to include a precision parameter for
consistency and accuracy.
[[1]](diffhunk://#diff-28a2e9e3b2d82e5dc455606ebd872ae2f9b5d972281b8b4c8716e780ddf31d23L77-R77)
[[2]](diffhunk://#diff-28a2e9e3b2d82e5dc455606ebd872ae2f9b5d972281b8b4c8716e780ddf31d23L87-R87)
[[3]](diffhunk://#diff-b51835bb15d2b09cefa96824aa9213239280d5ebcf401bcf97d7518d76788280L61-R61)
[[4]](diffhunk://#diff-b51835bb15d2b09cefa96824aa9213239280d5ebcf401bcf97d7518d76788280L100-R100)

### Utility Enhancements:
* Added an `is_one` function to the `BigDecimal` struct to check if a
number represents one.

These changes enhance the functionality and precision of the BigDecimal
library, particularly with the addition of the new root calculation
capabilities.